refactor: extract group member phases and roster presentation
