+++ apps/desktop/src/plugins/hermes-bots/group-rounds.ts
@@ -1,28 +1,25 @@
+import { botFriendlyNames, botHandle, mentionNameForms } from './data'
 /**
  * Room-level coordination: who speaks, in what order, for how long — the
  * @mention parse, the round-robin driver, the #93129 member holds, the stop
  * path, and the user send that starts it all.
  */
-
-import { botFriendlyNames, botHandle, clearBotAttention, mentionNameForms, noteBotAttention } from './data'
 import { recordGroupActivity } from './group-activity'
 import {
   $groupChats,
   $groupNeedsYou,
   appendGroupChatEntry,
-  GROUP_CHAT_HISTORY_LIMIT,
   GROUP_CHAT_MAX_CONTINUATIONS,
   GROUP_CHAT_MAX_MESSAGES,
   GROUP_CHAT_MAX_ROUNDS,
-  groupSpeakerLabel,
   groupThreadOf,
   mintGroupThreadId,
-  shouldCommitMemberTurn,
   updateGroupChat
 } from './group-chat'
 import type { GroupChatRoom, GroupHoldStamp } from './group-chat'
 import { durableGroupChatMembers, followGroupChat, groupMemberKey } from './group-membership'
-import { harvestStrandedGroupReply, isGroupPassText, runGroupChatMemberTurn } from './group-turns'
+import { runGroupContinuationMembers, runGroupRoundMember } from './group-round-members'
+import { harvestStrandedGroupReply } from './group-turns'
 import { requestForBot } from './routing'
 import type { Attachment, GroupMember, GroupMessage } from './types'
 
@@ -154,70 +151,6 @@ export function rotateGroupSpeakers(members: GroupMember[], round: number) {
   return [...members.slice(shift), ...members.slice(0, shift)]
 }
 
-/** Room-log line as a member sees it: `Name (user): …` / `Name: …` /
- *  `Name (you): …`. */
-export function formatGroupChatLine(entry: GroupMessage, viewerName: string) {
-  // Attachments are staged into each member's session as real payloads; the
-  // transcript line names them so the delta text and the bytes line up.
-  const attached =
-    Array.isArray(entry.images) && entry.images.length
-      ? ` ${entry.images
-          .map(img => {
-            const label = img.kind === 'pdf' ? 'attached PDF' : img.kind === 'file' ? 'attached file' : 'attached image'
-
-            return `[${label}: ${img.name || 'image'}]`
-          })
-          .join(' ')}`
-      : ''
-
-  if (entry.from.kind === 'user') {
-    return `${entry.from.name || 'User'} (user): ${entry.text}${attached}`
-  }
-
-  const suffix = entry.from.name === viewerName ? ' (you)' : ''
-  // Cross-connection speakers carry their device so same-named agents on
-  // two machines stay tellable apart in every member's transcript.
-  const source = entry.from.source ? ` [${entry.from.source}]` : ''
-
-  return `${groupSpeakerLabel(entry.from.name)}${suffix}${source}: ${entry.text}${attached}`
-}
-
-interface GroupChatTurnPromptInput {
-  deltaLines: string[]
-  groupName: string
-  members: GroupMember[]
-  viewer: GroupMember
-}
-
-/** The full per-turn payload for one member: participation rules + the room
- *  delta. Rules travel in the turn payload (not SOUL) so every existing bot
- *  can join a group chat without a profile migration. */
-export function buildGroupChatTurnPrompt({ groupName, members, viewer, deltaLines }: GroupChatTurnPromptInput) {
-  const viewerKey = groupMemberKey(viewer)
-  const peers = members.filter(m => groupMemberKey(m) !== viewerKey)
-
-  const peerNames = peers
-    .map(m => {
-      const handle = m.title ? `${m.title} (@${botHandle(m.name, m)})` : `@${botHandle(m.name, m)}`
-
-      return m.remoteSource ? `${handle} [on ${m.connectionLabel || m.connectionId}]` : handle
-    })
-    .join(', ')
-
-  return [
-    `[Group chat: "${groupName}"] You are @${botHandle(viewer.name, viewer)}, one participant in a group chat with ${peerNames || 'no one else yet'} and the user.`,
-    '',
-    'New messages in the room since your last turn (oldest first):',
-    ...deltaLines.map(line => `  ${line}`),
-    '',
-    'Rules for this room:',
-    '- Reply with ONE conversational message ONLY if you have something new worth adding: build on what was just said, claim or hand off work, answer a question aimed at you, or report a real result. Keep chatter short (1-3 sentences) — but when you are delivering a result, an answer the user asked for, or substantive work, give it at full quality and length; never thin out real content to fit the room.',
-    '- If you have nothing new to add, reply with exactly "(pass)". Passing is good — it lets the conversation settle.',
-    '- Mention a teammate as @name to pull them in; mention @user only for a judgment call or a result the user needs. Do not repeat points already made.',
-    '- Never reveal content from your private 1:1 chats. Your reply text goes to the room verbatim — no preamble, no meta-commentary.'
-  ].join('\n')
-}
-
 // --- member-hold helpers (#93129) — pure, unit-tested ---
 
 /** #93129: classify a USER room message's effect on member holds. Only user
@@ -323,13 +256,6 @@ export function applyGroupHoldDirective(
   }
 
   return next
-}
-
-/** #93129: a held member's skip must consume its delta exactly once —
- *  advance the watermark past the current log so the same entries never
- *  re-trigger the skip. Null = nothing to consume (no write, no spin). */
-export function heldMemberWatermarkAdvance(seen: number | undefined, logLength: number): null | number {
-  return logLength > (seen || 0) ? logLength : null
 }
 
 // --- end member-hold helpers ---
@@ -493,9 +419,24 @@ export async function stopGroupThread(group: string, thread: null | string, memb
  *  Watermarks are per thread+member (`${thread}::${memberKey}`), so parallel
  *  topics never eat each other's deltas. */
 export async function runGroupChatRounds(group: string, members: GroupMember[], thread: string) {
-  const binding = followGroupChat(group, name => { group = name })
+  const binding = followGroupChat(group, name => {
+    group = name
+  })
+
   const startEpoch = ($groupChats.get()[group] || {}).epoch || 0
   const isCurrent = () => binding.isLive() && (($groupChats.get()[group] || {}).epoch || 0) === startEpoch
+
+  const context = {
+    get group() {
+      return group
+    },
+    members,
+    thread,
+    startEpoch,
+    binding,
+    isCurrent
+  }
+
   let posted = 0
   let continuations = 0
   // #94478: how this drive ended. 'settled' means quiet consensus (everyone
@@ -565,183 +506,13 @@ export async function runGroupChatRounds(group: string, members: GroupMember[], 
           return
         }
 
-        const room = $groupChats.get()[group] || {
-          log: [],
-          watermarks: {}
-        }
+        const result = await runGroupRoundMember(context, member)
 
-        const memberKey = groupMemberKey(member)
-        const markKey = `${thread}::${memberKey}`
-        const seen = room.watermarks[markKey] || 0
-        // Delta: NEW room entries, narrowed to this thread — the member's
-        // turn sees only the conversation it's part of.
-        const delta = room.log.slice(seen).filter((e: GroupMessage) => groupThreadOf(e) === thread)
-
-        if (!delta.length) {
-          continue
-        }
-
-        // #93129: a member the user told to stop is HELD — no turn until an
-        // explicit release (resume / @all resume / a direct non-stop
-        // mention). Consume the delta exactly once (watermark past the
-        // current log) so the same entries never re-trigger this skip, and
-        // surface WHY the bot is silent in the activity feed the first time.
-        const heldEntry = (room.holds || {})[memberKey]
-
-        if (heldEntry) {
-          const advance = heldMemberWatermarkAdvance(seen, room.log.length)
-          updateGroupChat(group, (r: GroupChatRoom) => {
-            if (advance !== null) {
-              r.watermarks[markKey] = advance
-            }
-
-            if (r.holds?.[memberKey] && !r.holds[memberKey].noted) {
-              r.holds = {
-                ...r.holds,
-                [memberKey]: {
-                  ...r.holds[memberKey],
-                  noted: true
-                }
-              }
-            }
-
-            return r
-          })
-
-          if (!heldEntry.noted) {
-            recordGroupActivity(group, {
-              kind: 'held',
-              member: member.name,
-              thread
-            })
-          }
-
-          continue
-        }
-
-        const prompt = buildGroupChatTurnPrompt({
-          groupName: group,
-          members,
-          viewer: member,
-          deltaLines: delta
-            .slice(-GROUP_CHAT_HISTORY_LIMIT)
-            .map((e: GroupMessage) => formatGroupChatLine(e, member.name))
-        })
-
-        // Images riding this delta (user attachments — member entries don't
-        // carry images today, but flatMap keeps this future-proof) get staged
-        // into the member's session so the model sees the pixels, not just
-        // the transcript's [attached image: …] marker.
-        const deltaImages = delta.flatMap((e: GroupMessage) => (Array.isArray(e.images) ? e.images : []))
-
-        // Surface WHO is on turn (runtime-only, like running/epoch) so the
-        // room shows "Radar is thinking…" instead of a generic working line —
-        // long model turns otherwise read as the room being stuck.
-        updateGroupChat(group, (r: GroupChatRoom) => {
-          r.turn = member.name
-
-          return r
-        })
-        let reply: null | string = null
-
-        try {
-          reply = await runGroupChatMemberTurn(group, member, prompt, thread, deltaImages)
-
-          // Needs-attention hook (#93091 item 3): a turn that produced a real
-          // reply (or an explicit pass) is a good turn — clear the badge.
-          // A timed-out turn also returns null but never threw; leaving any
-          // prior badge in place there is the conservative choice.
-          if (reply !== null) {
-            clearBotAttention(groupMemberKey(member))
-          }
-        } catch (error: any) {
-          if (!binding.isLive()) {
-            return
-          }
-
-          const reason = String(error?.data?.reason || '').trim()
-          recordGroupActivity(group, {
-            kind: 'failed',
-            member: member.name,
-            thread,
-            ...(reason
-              ? {
-                  reason
-                }
-              : {})
-          })
-          noteBotAttention(groupMemberKey(member), reason || error?.message || error)
-          reply = null // a failed turn is a pass, never a room error
-        }
-
-        // #93127: the turn may have finished AFTER a newer user send bumped
-        // the room epoch. That newer send's loop re-drives this member with
-        // the full delta, so committing this stale result (watermark advance
-        // + append) would double-deliver the same reply. Drop it here —
-        // BEFORE the watermark advance and BEFORE the append. Only a newer
-        // USER entry in THIS thread makes the re-drive premise true: a
-        // cross-thread send bumps the epoch too, but its loop filters this
-        // thread out and would never regenerate the finished reply. The
-        // during-turn tail is anchored by entry id, not index — the history
-        // trim drops entries from the FRONT, so an index slice could
-        // overshoot after a mid-turn trim and silently commit a stale turn.
-        if (!binding.isLive()) {
+        if (!binding.isLive() || result === null) {
           return
         }
 
-        const roomNow = $groupChats.get()[group] || {
-          log: []
-        }
-
-        const epochNow = roomNow.epoch || 0
-        const anchorId = room.log.length ? room.log[room.log.length - 1].id : null
-        const anchorIdx = anchorId === null ? -1 : roomNow.log.findIndex((e: GroupMessage) => e.id === anchorId)
-        // Anchor trimmed away ⇒ every pre-turn entry was dropped, so every
-        // surviving entry is newer — scanning the whole log stays exact.
-        const turnTail = anchorIdx >= 0 ? roomNow.log.slice(anchorIdx + 1) : roomNow.log
-
-        const newerUserEntryInThread = turnTail.some(
-          (e: GroupMessage) => e.from?.kind === 'user' && groupThreadOf(e) === thread
-        )
-
-        if (!shouldCommitMemberTurn(startEpoch, epochNow, newerUserEntryInThread)) {
-          recordGroupActivity(group, {
-            kind: 'cancelled',
-            member: member.name,
-            thread
-          })
-
-          return
-        }
-
-        // The member has now seen everything up to the pre-reply log length.
-        updateGroupChat(group, (r: GroupChatRoom) => {
-          r.watermarks[markKey] = r.log.length
-
-          return r
-        })
-
-        if (reply !== null && !isGroupPassText(reply)) {
-          appendGroupChatEntry(
-            group,
-            {
-              kind: 'member',
-              name: member.name,
-              ...(member.remoteSource
-                ? {
-                    source: member.connectionLabel || member.connectionId
-                  }
-                : {})
-            },
-            reply,
-            thread
-          )
-          // Its own message counts as seen too.
-          updateGroupChat(group, (r: GroupChatRoom) => {
-            r.watermarks[markKey] = r.log.length
-
-            return r
-          })
+        if (result) {
           posted += 1
           spokeThisRound += 1
         }
@@ -762,125 +533,14 @@ export async function runGroupChatRounds(group: string, members: GroupMember[], 
         // room's entire budget on back-and-forth handoffs.
         continuations += 1
 
-        if (pendingKeys.length && continuations <= GROUP_CHAT_MAX_CONTINUATIONS) {
-          const citedMembers = members.filter((member: GroupMember) => pendingKeys.includes(groupMemberKey(member)))
+        const continued = await runGroupContinuationMembers(context, pendingKeys, continuations, posted)
 
-          if (citedMembers.length && posted < GROUP_CHAT_MAX_MESSAGES) {
-            const strandedNow = ($groupChats.get()[group] || {}).stranded || {}
-
-            const continuationResponders = citedMembers.filter(
-              (member: GroupMember) => !Object.prototype.hasOwnProperty.call(strandedNow, groupMemberKey(member))
-            )
-
-            for (const member of continuationResponders) {
-              if (!isCurrent() || posted >= GROUP_CHAT_MAX_MESSAGES || continuations > GROUP_CHAT_MAX_CONTINUATIONS) {
-                break
-              }
-
-              const room = $groupChats.get()[group] || {
-                log: [],
-                watermarks: {}
-              }
-
-              const memberKey = groupMemberKey(member)
-              const markKey = `${thread}::${memberKey}`
-              const seen = room.watermarks[markKey] || 0
-              const delta = room.log.slice(seen).filter((e: GroupMessage) => groupThreadOf(e) === thread)
-
-              // A cited member always has delta here (the citing reply IS in
-              // its tail); skip defensively anyway so an empty prompt never
-              // fires.
-              if (!delta.length) {
-                continue
-              }
-
-              const heldEntry = (room.holds || {})[memberKey]
-
-              if (heldEntry) {
-                continue // holds still apply to continuation turns (#93129)
-              }
-
-              const prompt = buildGroupChatTurnPrompt({
-                groupName: group,
-                members,
-                viewer: member,
-                // The continuation prompt centers on what the member missed:
-                // everything since its watermark, which includes the reply
-                // that cites it.
-                deltaLines: delta
-                  .slice(-GROUP_CHAT_HISTORY_LIMIT)
-                  .map((e: GroupMessage) => formatGroupChatLine(e, member.name))
-              })
-
-              updateGroupChat(group, (r: GroupChatRoom) => {
-                r.turn = member.name
-
-                return r
-              })
-              let continuationReply: null | string = null
-
-              try {
-                continuationReply = await runGroupChatMemberTurn(group, member, prompt, thread)
-
-                if (continuationReply !== null) {
-                  clearBotAttention(memberKey)
-                }
-              } catch (error: any) {
-                if (!binding.isLive()) {
-                  return
-                }
-
-                recordGroupActivity(group, {
-                  kind: 'failed',
-                  member: member.name,
-                  thread
-                })
-                noteBotAttention(memberKey, error?.message || error)
-                continuationReply = null
-              }
-
-              if (!isCurrent()) {
-                return
-              }
-
-              updateGroupChat(group, (r: GroupChatRoom) => {
-                r.watermarks[markKey] = r.log.length
-
-                return r
-              })
-
-              if (continuationReply !== null && !isGroupPassText(continuationReply)) {
-                appendGroupChatEntry(
-                  group,
-                  {
-                    kind: 'member',
-                    name: member.name,
-                    ...(member.remoteSource
-                      ? {
-                          source: member.connectionLabel || member.connectionId
-                        }
-                      : {})
-                  },
-                  continuationReply,
-                  thread
-                )
-                updateGroupChat(group, (r: GroupChatRoom) => {
-                  r.watermarks[markKey] = r.log.length
-
-                  return r
-                })
-                posted += 1
-
-                // The continuation's own reply may cite someone else — fall
-                // through to the normal loop so the next round handles it via
-                // the same responder machinery. Reaching here means the loop
-                // continues rather than settling; the outer for-loop's next
-                // iteration re-evaluates everything.
-                spokeThisRound += 1
-              }
-            }
-          }
+        if (!binding.isLive() || continued === null) {
+          return
         }
+
+        posted += continued
+        spokeThisRound += continued
 
         if (spokeThisRound === 0) {
           // Genuinely nothing left to say — including after the continuation
@@ -938,7 +598,9 @@ export async function runGroupChatRounds(group: string, members: GroupMember[], 
  *  stranded, a new loop takes the room over (it harvests on its own), or the
  *  room record disappears (disband). */
 async function harvestStrandedUntilSettled(group: string, members: GroupMember[], thread: string) {
-  const binding = followGroupChat(group, name => { group = name })
+  const binding = followGroupChat(group, name => {
+    group = name
+  })
 
   try {
     const HARVEST_INTERVAL_MS = 5000
@@ -1059,7 +721,9 @@ export function sendToGroupChat(
     thread: target
   })
 
-  const binding = followGroupChat(group, name => { group = name })
+  const binding = followGroupChat(group, name => {
+    group = name
+  })
 
   const drive = () => {
     if (!binding.isLive()) {
@@ -1068,15 +732,17 @@ export function sendToGroupChat(
       return
     }
 
-    void runGroupChatRounds(group, members, target).catch(() => {
-      if (binding.isLive()) {
-        updateGroupChat(group, (r: GroupChatRoom) => {
-          r.running = false
+    void runGroupChatRounds(group, members, target)
+      .catch(() => {
+        if (binding.isLive()) {
+          updateGroupChat(group, (r: GroupChatRoom) => {
+            r.running = false
 
-          return r
-        })
-      }
-    }).finally(binding.dispose)
+            return r
+          })
+        }
+      })
+      .finally(binding.dispose)
   }
 
   if (!wasRunning) {
